requirements.txt
numpy
Pillow
opensimplex
# plotly
# pandas
# pyglet
# matplotlib
# seaborn
pytest
mypy
toml
types-toml
# nimporter
setuptools
numba
